changed int to long

+++ 1. Very Easy/convertHoursIntoSeconds.cpp
@@ -9,7 +9,7 @@ https://edabit.com/challenge/y9Jhs33vdts5GPMW7
 
 #include<bits/stdc++.h>
 using namespace std;
-long howManySeconds(int hours){
+long howManySeconds(long hours){
 	return 60*60*hours;
 }
 int main(){
